Add tlsSkipVerify and authMethod flags
These replace the existing insecure and enableLoginAuth flags, which are
now deprecated. By combining the two new flags with the existing
smtpPort flag, users can replicate the behavior of the insecure flag,
but now have more options for supporting a wider range of SMTP servers.

Because of the interaction between these options, no longer use the
smtp.SendMail helper function, but implement the SMTP exchange using the
lower-level client methods.

Signed-off-by: Billy Keyes <bkeyes@palantir.com>